backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
networkx==3.1
matplotlib==3.7.2
seaborn==0.12.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic-settings==2.0.3
pymongo==4.5.0
redis==4.6.0
torch==2.0.1 # Or specify appropriate version/CPU/GPU variant
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1
gunicorn==21.2.0